Use dynamic batch size
The problem is when a model is run on 512 nodes * 8 GPUs * 12 batch
size = 24576 for global batch size. This old version of HuggingFace
will add this to the shape used by PyTorch symbolic_trace, resulting
in out-of-memory error on a 256GB node. This hack causes PyTorch to
trace using dynamic batch size. Then it fills in the graph with the
real values of the batch size. FlexFlow will later split it across
machines back to the expected smaller local batch size.